+++ lib/src/base/types/Dynamic.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import {formatName, sectionHeader, parseQS} from "../../utils";
+import {formatName, sectionHeader, parseQS, getGlobalTypeData} from "../../utils";
 
 export default {
   props: {
@@ -174,29 +174,20 @@ export default {
         instance_name: this.instanceName
       });
     },
-    getGlobalType() {
-      const token = this.$cookies.get("sections-auth-token");
-      const config = {
-        headers: sectionHeader({ token }),
-      };
-      const URL =
-          `${this.$sections.serverUrl}/project/${this.$sections.projectId}/global-instances/${this.props.linked_to}`;
+    async getGlobalType() {
       this.$emit("load", true);
-
-      this.$axios.get(URL, config).then((res) => {
-        if (res && res.data) {
-          this.$emit("load", false);
-          this.autoInsert = res.data.auto_insertion
-          if (res.data.pages && res.data.pages.length > 0) {
-            this.pages = res.data.pages.map(p => p.path)
-          }
-          this.instanceName = res.data.name
+      const result = await getGlobalTypeData(this.props.linked_to)
+      if (result.res && result.res.data) {
+        this.$emit("load", false);
+        this.autoInsert = result.res.data.auto_insertion
+        if (result.res.data.pages && result.res.data.pages.length > 0) {
+          this.pages = result.res.data.pages.map(p => p.path)
         }
-      })
-          .catch((error) => {
-            this.$emit("load", false);
-            this.showToast("Error", "error", error.response.data.message, error.response.data.options);
-          });
+        this.instanceName = result.res.data.name
+      } else if (result.error) {
+        this.$emit("load", false);
+        this.showToast("Error", "error", result.error.response.data.message, result.error.response.data.options);
+      }
     }
   },
 };